Only stop cassandra after tests if the cassandra.pid file exists.
Also, decrease polling the file system in ReplyTS a bit.

Reviewed-on: https://gerrit.ecg.so/33060
Tested-by: Gerrit-At-Gitmaster001

+++ bin/build.sh
@@ -32,10 +32,12 @@ function startCassandra() {
 }
 
 function stopCassandra() {
-    log "Stopping cassandra"
-    kill $(cat $CASSANDRA_PID)
-    rm ${CASSANDRA_PID}
-    rm -rf ${CASSANDRA_DIR}
+    if [[ -e ${CASSANDRA_PID} ]]; then
+        log "Stopping cassandra"
+        kill $(cat ${CASSANDRA_PID})
+        rm ${CASSANDRA_PID}
+        rm -rf ${CASSANDRA_DIR}
+    fi
 }
 
 function parseCmd() {
